Add support for JSON field called Device
Added support for JSON field called "Device" to specify which device to cast to.  Goes along with modification to mdns code to look at the text name of the device (not the network name).  ie Kitchen rather than Google-Home-blah-blah-blah.  if you don't specify a device, it uses the 'default' one listed at the top of the code.  Note that error handling on bad device name is wonky.  can't really trap that very well and so it end up stuck in the queue.